requirements.txt
avro==1.10.2
kafka-python==2.0.2
